move to Dapper.AOT.Test.Integration.csproj
