+++ test/Dapper.AOT.Test.Integration/BatchTests.cs
@@ -5,6 +5,7 @@ using System.Collections.Immutable;
 using System.Collections.ObjectModel;
 using System.Data;
 using System.Linq;
+using Dapper.AOT.Test.Integration.Setup;
 using Xunit;
 
 namespace Dapper.AOT.Test.Integration;
