+++ test/Dapper.AOT.Test.Integration/BulkInsertIngegrationTests.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Dapper.AOT.Test.Integration.Setup;
 using Xunit;
 
 namespace Dapper.AOT.Test.Integration;
