+++ test/Dapper.AOT.Test.Integration/DynamicTests.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Data.SqlClient;
 using System;
 using System.Collections.Generic;
+using Dapper.AOT.Test.Integration.Setup;
 using Xunit;
 
 namespace Dapper.AOT.Test.Integration;
